# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cashtree Revamp API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/home": {
            "get": {
                "tags": [
                    "home"
                ],
                "summary": "Fetching data home",
                "description": "Fetching data",
                "operationId": "fetchingdata",
                "parameters": [
                    {
                        "name": "mmses",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "000"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/notification": {
            "get": {
                "tags": [
                    "notification"
                ],
                "summary": "list of notification",
                "operationId": "App\\Http\\Controllers\\Api\\NotificationController::index",
                "parameters": [
                    {
                        "name": "mmses",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of notifications"
                    }
                }
            }
        },
        "/api/notification/read": {
            "post": {
                "tags": [
                    "notification"
                ],
                "summary": "list of notification",
                "operationId": "App\\Http\\Controllers\\Api\\NotificationController::read",
                "parameters": [
                    {
                        "name": "mmses",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "notification_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success message"
                    },
                    "201": {
                        "description": "Notification ID Mandatory"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}